PHPLIB-1531: Replace FailPointObserver with different logic (#1427)
* Extract fail point management to a trait used by Context

* Ensure event listeners are still unregistered after an operation fails

This change was prompted by the addition of a finally block after executing operations. It may address a memory leak whereby listeners (and the Context they reference) would remain in PHPC's global registry until RSHUTDOWN.

---------
